add simulate service and j unit

+++ transaction-service/src/main/java/com/transaction/app/dto/insight/InsightResponse.java
@@ -12,5 +12,5 @@ public class InsightResponse {
     private String insightMessage;
     private Double futureValue;
     private Double monthlyInvestmentNeeded;
-    private Long monthsToAchieve;
+    private int monthsToAchieve;
 }
